editValues folder: Fix describe description and rename file
