+++ test/unit/operations/write/editValues/editLeafValuesOffARoot.ts
@@ -1,5 +1,6 @@
 import { CacheContext } from '../../../../../src/context';
 import { GraphSnapshot } from '../../../../../src/GraphSnapshot';
+import { EntitySnapshot } from '../../../../../src/nodes';
 import { write } from '../../../../../src/operations/write';
 import { NodeId, StaticNodeId } from '../../../../../src/schema';
 import { query, strictConfig } from '../../../../helpers';
@@ -16,40 +17,19 @@ describe(`operations.write`, () => {
   const empty = new GraphSnapshot();
   const valuesQuery = query(`{ foo bar }`);
 
-  describe(`edit nested values of a root`, () => {
+  describe(`edit leaf-values hanging off of a root`, () => {
     let baseline: GraphSnapshot, snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
     beforeAll(() => {
-      const baselineResult = write(context, empty, valuesQuery, {
-        foo: [{ value: 1 }, { value: 2 }, { value: 3 }],
-        bar: { baz: 'asdf' },
-      });
+      const baselineResult = write(context, empty, valuesQuery, { foo: 123, bar: { baz: 'asdf' } });
       baseline = baselineResult.snapshot;
 
-      const result = write(context, baseline, valuesQuery, {
-        foo: [{ value: -1 }, { extra: true }],
-        bar: {
-          baz: 'fdsa',
-          fizz: 'buzz',
-        },
-      });
+      const result = write(context, baseline, query(`{ foo }`), { foo: 321 });
       snapshot = result.snapshot;
       editedNodeIds = result.editedNodeIds;
     });
 
-    it(`merges new properties with existing objects`, () => {
-      expect(snapshot.getNodeData(QueryRootId).bar).to.deep.eq({ baz: 'fdsa', fizz: 'buzz' });
-    });
-
-    it(`honors array lengths`, () => {
-      expect(snapshot.getNodeData(QueryRootId).foo.length).to.eq(2);
-    });
-
-    it(`overwrites previous values in array elements`, () => {
-      expect(snapshot.getNodeData(QueryRootId).foo[0]).to.deep.eq({ value: -1 });
-    });
-
-    it(`no merging of new values in array elements as we copy leaf value`, () => {
-      expect(snapshot.getNodeData(QueryRootId).foo[1]).to.deep.eq({ extra: true });
+    it(`updates the value, and its container`, () => {
+      expect(snapshot.getNodeData(QueryRootId)).to.deep.eq({ foo: 321, bar: { baz: 'asdf' } });
     });
 
     it(`marks the root as edited`, () => {
@@ -60,6 +40,9 @@ describe(`operations.write`, () => {
       expect(snapshot.allNodeIds()).to.have.members([QueryRootId]);
     });
 
-  });
+    it(`emits the edited node as an EntitySnapshot`, () => {
+      expect(snapshot.getNodeSnapshot(QueryRootId)).to.be.an.instanceOf(EntitySnapshot);
+    });
 
+  });
 });
