+++ test/unit/operations/write/editValues/editReferencesInAnArray.ts
@@ -224,5 +224,4 @@ describe(`operations.write`, () => {
     });
 
   });
-
 });
